requirements.txt
pytesseract
Pillow
PyPDF2